Update dependencies; Use kotlin 1.3.0 and coroutines 1.0.0;

+++ screensadapter-compiler/src/main/java/com/e16din/screensadapter/processor/ScreensAdapterProcessor.kt
@@ -304,7 +304,7 @@ class ScreensAdapterProcessor : AbstractProcessor() {
                         .addParameter(androidAppParamName, CLS_APPLICATION)
                         .addParameter(appParamName, ClassName.bestGuess(appClassName!!))
                         .addParameter(serverParamName, ClassName.bestGuess(serverClassName!!))
-                        .addParameter(delayForSplashMsParamName, Int::class)
+                        .addParameter(delayForSplashMsParamName, Long::class)
                         .build())
                 .addSuperclassConstructorParameter(androidAppParamName)
                 .addSuperclassConstructorParameter(appParamName)
